+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/ui/detail/RepositoryDetailScreen.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codecheck.ui
+package jp.co.yumemi.android.codecheck.ui.detail
 
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.core.screen.Screen
